% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEigenVectBestK.R
\name{computeEigenVectBestK}
\alias{computeEigenVectBestK}
\title{computeEigenVectBestK}
\usage{
computeEigenVectBestK(ioSubEnv)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
Compute the best number of eigen vectors using a two regression fit
}

