% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEigenVectVal.R
\name{computeEigenVectVal}
\alias{computeEigenVectVal}
\title{computeEigenVectVal}
\usage{
computeEigenVectVal(ioSubEnv)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
computeEigenVectVal computes the eigen vectors and eigen values from the laplacian matrix.
Only a percentage of eigen vectors is kepts (either following input arguments or automatically
with an elbow-like heuristic)
}

