% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubInputData.random.R
\name{createSubInputData.random}
\alias{createSubInputData.random}
\title{createSubInputData.random}
\usage{
createSubInputData.random(ioSubEnv, outDirPath)
}
\arguments{
\item{ioSubEnv}{a global environment variable}

\item{outDirPath}{output directory path where all ScaleNet outputs are saved}
}
\description{
Create input sub datasets : random
}

