% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubInputData.R
\name{createSubInputData}
\alias{createSubInputData}
\title{createSubInputData}
\usage{
createSubInputData(ioSubEnv)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
createSubInputData
}

