% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubInputData.spectral.R
\name{createSubInputData.spectral}
\alias{createSubInputData.spectral}
\title{createSubInputData.spectral}
\usage{
createSubInputData.spectral(ioSubEnv, outDirPath)
}
\arguments{
\item{ioSubEnv}{a global environment variable}

\item{outDirPath}{output directory path where all scalenet outputs are saved}
}
\description{
Create input sub datasets : spectral
}
