% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubInputData.spectralBipartition.R
\name{createSubInputData.spectralBipartition}
\alias{createSubInputData.spectralBipartition}
\title{createSubInputData.spectralBipartition}
\usage{
createSubInputData.spectralBipartition(ioSubEnv, outDirPath)
}
\arguments{
\item{ioSubEnv}{a global environment variable}

\item{outDirPath}{output directory path where all scalenet outputs are saved}
}
\description{
Create input sub datasets : spectral bi partitionning
}
