% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubInputData.spectralFuzzyCmeansOrder.R
\name{createSubInputData.spectralFuzzyCmeansOrder}
\alias{createSubInputData.spectralFuzzyCmeansOrder}
\title{createSubInputData.spectralFuzzyCmeansOrder}
\usage{
createSubInputData.spectralFuzzyCmeansOrder(ioSubEnv, outDirPath)
}
\arguments{
\item{ioSubEnv}{a global environment variable}

\item{outDirPath}{output directory path where all scalenet outputs are saved}
}
\description{
Create input sub datasets : spectral fuzzy cmeans (order)
- make 2*subset.k fuzzy clusters
- for each fuzzy cluster, keep only the m variables with the strongest membership
}
