% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubInputData.spectralFuzzyCmeansSample.R
\name{createSubInputData.spectralFuzzyCmeansSample}
\alias{createSubInputData.spectralFuzzyCmeansSample}
\title{createSubInputData.spectralFuzzyCmeansSample}
\usage{
createSubInputData.spectralFuzzyCmeansSample(ioSubEnv, outDirPath)
}
\arguments{
\item{ioSubEnv}{a global environment variable}

\item{outDirPath}{output directory path where all ScaleNet outputs are saved}
}
\description{
Create input sub datasets : spectral fuzzy cmeans (sample)
- make 1/m clusters
- assign each variable to a cluster based on its membership
(use 'sample' based on the membership as probabilities and perform 20 assignments)
}
