% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubInputData.spectralKmeans.R
\name{createSubInputData.spectralKmeans}
\alias{createSubInputData.spectralKmeans}
\title{createSubInputData.spectralKmeans}
\usage{
createSubInputData.spectralKmeans(ioSubEnv, outDirPath)
}
\arguments{
\item{ioSubEnv}{a global environment variable}

\item{outDirPath}{output directory path where all ScaleNet outputs are saved}
}
\description{
Create input sub datasets : spectral K-means
}

