% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubOutputGraphs.R
\name{createSubOutputGraphs}
\alias{createSubOutputGraphs}
\title{createSubInputData}
\usage{
createSubOutputGraphs(ioSubEnv)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
createSubInputData
}
