% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{discretize}
\usage{
discretize(argInData, argDisc = "pamk", argBins = 5, argMaxClusters = 5,
  argVerbose = FALSE)
}
\arguments{
\item{argInData}{input data file path (a tab separated file) or a data.frame, with features in columns and observations in rows}

\item{argDisc, }{the discretizing method, "pamk", "equalfreq" or "equalwidth"; default is "pamk"}

\item{argBins, }{the required number of bins (used with "equalfreq" or "equalwidth"; default is 5)}

\item{argMaxClusters, }{the expected maximum number of clusters (only used with the "pamk" method)}
}
\description{
discretize discretizes numerical column of a data.frame. The data.frame can contain NA values.
Outlier values are identified using the Tukey's test. A pamk clustering is done on the non NA/no outlier values.
Upper outlier values are inserted in the cluster with the largest medoid, and lower outlier values with
the cluster of smallest medoid. If non outlier values are all equal, the values are put in one cluster, and the outlier
values, if any, are put in different clusters.
}
