% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaplacianRW.R
\name{LaplacianRW}
\alias{LaplacianRW}
\title{LaplacianRW}
\usage{
LaplacianRW(ioSubEnv)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
LaplacianRW computes the LAPLACIAN MATRIX, Lrw=D^-1.L=I-D^-1.W, where W is the
mutual information matrix, D=diag(di) and di=sum(Coli)
}

