% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSaveEigenVal.R
\name{plotSaveEigenVal}
\alias{plotSaveEigenVal}
\title{plotSaveEigenVal}
\usage{
plotSaveEigenVal(ioSubEnv)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
plotSaveEigenVal
}

