% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSaveEigenVect.R
\name{plotSaveEigenVect}
\alias{plotSaveEigenVect}
\title{plotSaveEigenVect}
\usage{
plotSaveEigenVect(ioSubEnv)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
plotSaveEigenVect
}

