% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethod.aracne.R
\name{rMethod.aracne}
\alias{rMethod.aracne}
\title{rMethod.aracne}
\usage{
rMethod.aracne(argInData, argOutDir, argEstimator = "mi.mm", argEpsilon = 0,
  argVerbose = FALSE)
}
\arguments{
\item{argInData}{input data file, a tab separated file with features in columns and observations in rows}

\item{argOutDir}{output directory file, will be created}

\item{argEstimator}{mutual information estimator}

\item{argEpsilon}{constraint for third edge removal}

\item{argVerbose}{display details}
}
\description{
rMethod.aracne reconstruct network using the aracne approach (from minet package)
}

