% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethod.hc.R
\name{rMethod.hc}
\alias{rMethod.hc}
\title{rMethod.hc}
\usage{
rMethod.hc(argInData, argOutDir, argScore = "bde", argRestart = 20,
  argVerbose = FALSE, maxIter = 10000)
}
\arguments{
\item{argInData}{input data file, a tab separated file with features in columns and observations in rows}

\item{argOutDir}{output directory file, will be created}

\item{argScore}{scoring function}

\item{argRestart}{search and score restart}

\item{argVerbose}{display details}
}
\description{
rMethod.hc reconstruct network using the hhill climbing approach (from bnlearn package)
}

