% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleNet.convertToScaleNetFormat.R
\name{scaleNet.convertToScaleNetFormat}
\alias{scaleNet.convertToScaleNetFormat}
\title{scaleNet.convertToScaleNetFormat}
\usage{
scaleNet.convertToScaleNetFormat(ioSubEnv, iVarSpl)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
Convert into a scaleNet edgeList
}

