% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleNet.gatherSubGraphs.R
\name{scaleNet.gatherSubGraphs}
\alias{scaleNet.gatherSubGraphs}
\title{scaleNet.gatherSubGraphs}
\usage{
scaleNet.gatherSubGraphs(ioSubEnv, iVarSpl)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
Gather the sub graphs
- if the input datasets are based on spectralFuzzyCmeansSample
- save a globalNet for each assignment (up to ioSubEnv$nbVarSample)
}

