% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalenet.R
\name{scalenet}
\alias{scalenet}
\title{scalenet}
\usage{
scalenet(argInData, argOutDir, argEigenPerc = -1, argVarPerc,
  argSubsetType = "spectral", argReconsMeth = "bayes_hc",
  argReconsParam = list(bayes_hc = list(score = "bde", restart = 20)),
  argPresFreqThresh = c(0.5, 1), argNbSamples = 50000, argNumSeed = 6196,
  argNbCPU = 2, argDiscretize = FALSE, argVerbose = FALSE)
}
\arguments{
\item{argInData}{input data file path (a tab separated file) or a data.frame, with features in columns and observations in rows}

\item{argOutDir}{output directory file, will be created}

\item{argEigenPerc}{percentage of eigen vectors to use,
if -1 this percentage is set with an internal elbow-like heuristic method}

\item{argSubsetType}{variable subset selection, subset of variables can be selected from:
(1) eigen vector elements ("spectral"), (2) spectral k-means clusters ("spectralKmeans"),
(3) spectral fuzzy c-means clusters ("spectralFuzzyCmeansOrder" or "spectralFuzzyCmeansSample")
(4) spectral bipartition clusters ("spectralBipartition" )}

\item{argReconsMeth}{reconstruction method, (1) bayesian hill climbing ("bayes_hc"),
(2) aracne ("aracne")}

\item{argReconsParam}{reconstruction method parameters, (1) [bayes_hc] "-s score -r restartNumber",
(2) [aracne] "-p epsilon -w estimator"}

\item{argNbSamples}{number of observation to sample from the original dataset, if more than the
original number of observations, the maximum number of available samples is considered}

\item{argNumSeed}{random seed to sample observations from the original dataset}

\item{argNbCPU}{number of CPU to be used for local recontructions}

\item{argVerbose}{display details}
}
\description{
scalenet reconstruct network from local prediction based on subset of variables selected
with different clustering methods. The default subset selection correspond to the core
scaleNet approach. The other subset selection methods are left for comparison.
}
