% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleNet.spectralBipartition.R
\name{scaleNet.spectralBipartition}
\alias{scaleNet.spectralBipartition}
\title{scaleNet.spectralBipartition}
\usage{
scaleNet.spectralBipartition(ioSubEnv, sub.varNames)
}
\arguments{
\item{ioSubEnv}{a global environment variable}

\item{outDirPath}{output directory path where all ScaleNet outputs are saved}
}
\description{
Make clustering using a recursive bi-partitionning approach based on the 2nd eigenvector
}

