% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scs.R
\name{scs}
\alias{scs}
\title{scs}
\usage{
scs(scaleNetDir, argInData, argReconsMeth = c("aracne", "bayes_hc"),
  argReconsMethInfo = list(aracne = list(ort = "n", eweight =
  "epresenceScore"), bayes_hc = list(ort = "y", eweight = "ecorr")),
  argReconsParam = list(aracne = list(estimator = "mi.mm", epsilon = 0.001),
  bayes_hc = list(score = "bde", restart = 21), varPerc = 0.2),
  argPresFreqThresh = c(0.5, 1), argVerbose = FALSE)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
Create a consensus from scaleNet embedded method results
}

