% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEnvironment.R
\name{setEnvironment}
\alias{setEnvironment}
\title{setEnvironment}
\usage{
setEnvironment(
  inData,
  outDir,
  eigenPerc,
  varPerc,
  subsetType,
  reconsMeth,
  reconsParam,
  presFreqThresh,
  nbSamples,
  numSeed,
  nbCPU,
  disc,
  verbose,
  ioSubEnv
)
}
\arguments{
\item{inData}{input data file path (a tab separated file) or a data.frame, with features in columns and observations in rows}

\item{outDir}{output directory file, will be created}

\item{eigenPerc}{percentage of eigen vectors to use,
if -1 this percentage is set with an internal elbow-like heuristic method}

\item{subsetType}{variable subset selection, subset of variables can be selected from:
(1) eigen vector elements ("spectral"),
(2) spectral k-means clusters ("spectralKmeans"),
(3) spectral fuzzy c-means clusters ("spectralFuzzyCmeansOrder" or "spectralFuzzyCmeansSample")
(4) spectral bipartition clusters ("spectralBipartition" )}

\item{reconsMeth}{reconstruction method,
(1) bayesian hill climbing ("bayes_hc"),
(2) aracne ("aracne")}

\item{reconsParam}{reconstruction method parameters,
(1) [bayes_hc] "-s score -r restartNumber",
(2) [aracne] "-p epsilon -w estimator"}

\item{nbSamples}{number of observation to sample from the original dataset, if more than the
original number of observations, the maximum number of available samples is considered}

\item{numSeed}{random seed to sample observations from the original dataset}

\item{nbCPU}{number of CPU to be used for local recontructions}

\item{verbose}{display details}

\item{ioSubEnv}{a global environment variable}
}
\description{
setEnvironment sets all required information in en env() variables for reconstruction
}
