% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleNet.convertToScaleNetFormat.R
\name{scaleNet.convertToscalenetFormat}
\alias{scaleNet.convertToscalenetFormat}
\title{scaleNet.convertToscalenetFormat}
\usage{
scaleNet.convertToscalenetFormat(ioSubEnv, iVarSpl)
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
Convert into a scaleNet edgeList
}
