% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCS.R
\name{SCS}
\alias{SCS}
\title{SCS}
\usage{
SCS(scaleNetDir, argInData, argReconsMeth = c("aracne", "bayes_hc"),
  argReconsMethInfo = list(aracne = list(ort = "n", eweight =
  "epresenceScore"), bayes_hc = list(ort = "y", eweight = "ecorr")),
  argReconsParam = NULL, argPresFreqThresh = c(0.5, 1))
}
\arguments{
\item{ioSubEnv}{a global environment variable}
}
\description{
Create a consensus from scaleNet embedded method results
}

