% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scs.R
\name{scs}
\alias{scs}
\title{scs}
\usage{
scs(workspaceDir, argInData, argReconsMeth = c("aracne", "bayes_hc"),
  argReconsMethInfo = list(aracne = list(ort = "n", eweight =
  "epresenceScore"), bayes_hc = list(ort = "y", eweight = "ecorr")),
  argEmbReconsParam = list(aracne = list(estimator = "mi.mm", epsilon =
  0.001), bayes_hc = list(score = "bde", restart = 21), varPerc = 0.2),
  argPresFreqThresh = c(0.5, 1), clean.workspace = TRUE,
  argDiscretize = FALSE, argVerbose = FALSE)
}
\arguments{
\item{workspaceDir}{output directory file path (to be created)}

\item{argInData}{input data file path (a tab separated file) or a data.frame, with features in columns and observations in rows}

\item{argReconsMeth}{string vector with reconstruction method name}

\item{argReconsMethInfo}{list of complementary information on each reconstruction method. The
provided information are: "ort" (the method provides orientation n/y) and edge weight "eweight"
(epresenceScore/ecorr)}

\item{argEmbReconsParam}{list specific parameters for each reconstruction method and complementary parameters
specific to scalenet, namely: "nbSamples", "numSeed", "nbCPU", "eigenPerc", "varPerc" and "subsetType"}

\item{argPresFreqThresh}{a numeric vector of threshold for the frequency of edges from all subgraphs}

\item{clean.workspace}{if FALSE, all output files are kept}

\item{argVerbose}{display details}
}
\description{
Create a consensus from scaleNet embedded method results
}
